This is quite convoluted.
We have problems handling resizes of > 20K pix (w/h) images, given our memory constraints
The Adlib API does not have an API to resize images for a max width/height

so...

Restrict the images to max 10K pix (w/h)
Download the full size image to determine its size and aspect ratio
Generate an Adlib URL to resize the image to the wanted w/h and format

That means that for all images we download the image twice.
At least we don't burden our server with 120MB JPEGs.